+++ app/src/langs/es/es.lang.ts
@@ -27,5 +27,12 @@ export default {
         kelvin: 'Kº',
       },
     },
+    weight: {
+      title: 'Peso',
+      suffix: {
+        kgs: 'kgs',
+        lbs: 'lbs',
+      },
+    },
   },
 };