+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Created weight component.
+
+## 2022-08-06 (v0.0.10)
+
+### Added
+
 - Markdown documents are ignored on validation/deployment.
 
 ### Modified